+++ src/test/java/net/sf/jabref/logic/importer/fetcher/IsbnViaChimboriFetcherTest.java
@@ -16,35 +16,37 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotEquals;
 
 @Category(FetcherTests.class)
-public class IsbnViaChimboriFetcherTest {
-
-    private IsbnViaChimboriFetcher fetcher;
-    private BibEntry bibEntry;
+public class IsbnViaChimboriFetcherTest extends AbstractIsbnFetcherTest {
 
     @Before
     public void setUp() {
-        fetcher = new IsbnViaChimboriFetcher(JabRefPreferences.getInstance().getImportFormatPreferences());
-
         bibEntry = new BibEntry();
         bibEntry.setType(BibLatexEntryTypes.BOOK);
+        bibEntry.setField("bibtexkey", "9780321356680");
         bibEntry.setField("title", "Effective Java (Java Series)");
         bibEntry.setField("publisher", "Addison-Wesley Professional");
         bibEntry.setField("year", "2008");
         bibEntry.setField("author", "Joshua Bloch");
+        bibEntry.setField("isbn", "978-0321356680");
         bibEntry.setField("url", "https://www.amazon.com/Effective-Java-Joshua-Bloch-ebook/dp/B00B8V09HY%3FSubscriptionId%3D0JYN1NVW651KCA56C102%26tag%3Dtechkie-20%26linkCode%3Dxm2%26camp%3D2025%26creative%3D165953%26creativeASIN%3DB00B8V09HY");
+
+        fetcher = new IsbnViaChimboriFetcher(JabRefPreferences.getInstance().getImportFormatPreferences());
     }
 
     @Test
+    @Override
     public void testName() {
-        assertEquals("ISBN", fetcher.getName());
+        assertEquals("ISBN (Chimbori/Amazon)", fetcher.getName());
     }
 
     @Test
+    @Override
     public void testHelpPage() {
         assertEquals("ISBNtoBibTeX", fetcher.getHelpPage().getPageName());
     }
 
     @Test
+    @Override
     public void searchByIdSuccessfulWithShortISBN() throws FetcherException {
         Optional<BibEntry> fetchedEntry = fetcher.performSearchById("0321356683");
         bibEntry.setField("bibtexkey", "0321356683");
@@ -53,32 +55,20 @@ public class IsbnViaChimboriFetcherTest {
     }
 
     @Test
-    public void searchByIdSuccessfulWithLongISBN() throws FetcherException {
-        Optional<BibEntry> fetchedEntry = fetcher.performSearchById("978-0321356680");
-        bibEntry.setField("bibtexkey", "9780321356680");
-        bibEntry.setField("isbn", "978-0321356680");
+    @Override
+    public void authorsAreCorrectlyFormatted() throws Exception {
+        BibEntry bibEntry = new BibEntry();
+        bibEntry.setType(BibLatexEntryTypes.BOOK);
+        bibEntry.setField("bibtexkey", "3642434738");
+        bibEntry.setField("title", "Fundamentals of Business Process Management");
+        bibEntry.setField("publisher", "Springer");
+        bibEntry.setField("year", "2015");
+        bibEntry.setField("author", "Marlon Dumas and Marcello La Rosa and Jan Mendling and Hajo Reijers");
+        bibEntry.setField("isbn", "3642434738");
+        bibEntry.setField("url", "https://www.amazon.com/Fundamentals-Business-Process-Management-Marlon/dp/3642434738%3FSubscriptionId%3D0JYN1NVW651KCA56C102%26tag%3Dtechkie-20%26linkCode%3Dxm2%26camp%3D2025%26creative%3D165953%26creativeASIN%3D3642434738");
+
+        Optional<BibEntry> fetchedEntry = fetcher.performSearchById("3642434738");
         assertEquals(Optional.of(bibEntry), fetchedEntry);
-    }
-
-    @Test
-    public void searchByIdReturnsEmptyWithEmptyISBN() throws FetcherException {
-        Optional<BibEntry> fetchedEntry = fetcher.performSearchById("");
-        assertEquals(Optional.empty(), fetchedEntry);
-    }
-
-    @Test(expected = FetcherException.class)
-    public void searchByIdThrowsExceptionForShortInvalidISBN() throws FetcherException {
-        fetcher.performSearchById("123456789");
-    }
-
-    @Test(expected = FetcherException.class)
-    public void searchByIdThrowsExceptionForLongInvalidISB() throws FetcherException {
-        fetcher.performSearchById("012345678910");
-    }
-
-    @Test(expected = FetcherException.class)
-    public void searchByIdThrowsExceptionForInvalidISBN() throws FetcherException {
-        fetcher.performSearchById("jabref-4-ever");
     }
 
     @Test